panchanga/requirements.txt
pdf2image>=1.17.0
pytesseract>=0.3.10
Pillow>=10.0.0
pandas>=2.0.0
icalendar>=5.0.0
